libs/net: Move into internal
