+++ abci/client/grpc_client_test.go
@@ -16,8 +16,8 @@ import (
 
 	"google.golang.org/grpc/credentials/insecure"
 
+	cmtnet "github.com/cometbft/cometbft/internal/net"
 	"github.com/cometbft/cometbft/libs/log"
-	cmtnet "github.com/cometbft/cometbft/libs/net"
 
 	abciserver "github.com/cometbft/cometbft/abci/server"
 	"github.com/cometbft/cometbft/abci/types"
