+++ abci/server/grpc_server.go
@@ -7,7 +7,7 @@ import (
 	"google.golang.org/grpc"
 
 	"github.com/cometbft/cometbft/abci/types"
-	cmtnet "github.com/cometbft/cometbft/libs/net"
+	cmtnet "github.com/cometbft/cometbft/internal/net"
 	"github.com/cometbft/cometbft/libs/service"
 )
