+++ abci/server/socket_server.go
@@ -11,8 +11,8 @@ import (
 	"runtime"
 
 	"github.com/cometbft/cometbft/abci/types"
+	cmtnet "github.com/cometbft/cometbft/internal/net"
 	cmtlog "github.com/cometbft/cometbft/libs/log"
-	cmtnet "github.com/cometbft/cometbft/libs/net"
 	"github.com/cometbft/cometbft/libs/service"
 	cmtsync "github.com/cometbft/cometbft/libs/sync"
 )
@@ -162,7 +162,7 @@ func (s *SocketServer) waitForClose(closeConn chan error, connID int) {
 // Read requests from conn and deal with them
 func (s *SocketServer) handleRequests(closeConn chan error, conn io.Reader, responses chan<- *types.Response) {
 	var count int
-	var bufReader = bufio.NewReader(conn)
+	bufReader := bufio.NewReader(conn)
 
 	defer func() {
 		// make sure to recover from any app-related panics to allow proper socket cleanup.
@@ -184,7 +184,7 @@ func (s *SocketServer) handleRequests(closeConn chan error, conn io.Reader, resp
 
 	for {
 
-		var req = &types.Request{}
+		req := &types.Request{}
 		err := types.ReadMessage(bufReader, req)
 		if err != nil {
 			if err == io.EOF {
@@ -308,9 +308,9 @@ func (s *SocketServer) handleRequest(ctx context.Context, req *types.Request) (*
 // Pull responses from 'responses' and write them to conn.
 func (s *SocketServer) handleResponses(closeConn chan error, conn io.Writer, responses <-chan *types.Response) {
 	var count int
-	var bufWriter = bufio.NewWriter(conn)
+	bufWriter := bufio.NewWriter(conn)
 	for {
-		var res = <-responses
+		res := <-responses
 		err := types.WriteMessage(res, bufWriter)
 		if err != nil {
 			closeConn <- fmt.Errorf("error writing message: %w", err)
