+++ abci/tests/benchmarks/parallel/parallel.go
@@ -6,11 +6,10 @@ import (
 	"log"
 
 	"github.com/cometbft/cometbft/abci/types"
-	cmtnet "github.com/cometbft/cometbft/libs/net"
+	cmtnet "github.com/cometbft/cometbft/internal/net"
 )
 
 func main() {
-
 	conn, err := cmtnet.Connect("unix://test.sock")
 	if err != nil {
 		log.Fatal(err.Error())
@@ -20,7 +19,7 @@ func main() {
 	go func() {
 		counter := 0
 		for {
-			var res = &types.Response{}
+			res := &types.Response{}
 			err := types.ReadMessage(conn, res)
 			if err != nil {
 				log.Fatal(err.Error())
@@ -35,8 +34,8 @@ func main() {
 	// Write a bunch of requests
 	counter := 0
 	for i := 0; ; i++ {
-		var bufWriter = bufio.NewWriter(conn)
-		var req = types.ToRequestEcho("foobar")
+		bufWriter := bufio.NewWriter(conn)
+		req := types.ToRequestEcho("foobar")
 
 		err := types.WriteMessage(req, bufWriter)
 		if err != nil {
