+++ privval/utils.go
@@ -6,8 +6,8 @@ import (
 	"net"
 
 	"github.com/cometbft/cometbft/crypto/ed25519"
+	cmtnet "github.com/cometbft/cometbft/internal/net"
 	"github.com/cometbft/cometbft/libs/log"
-	cmtnet "github.com/cometbft/cometbft/libs/net"
 )
 
 // IsConnTimeout returns a boolean indicating whether the error is known to
